Break Relation out into its own component.
