+++ src/components/Relation/index.js
@@ -1,0 +1,5 @@
+import ilComponent from './Relation'
+
+export default Vue => {
+    Vue.component(ilComponent.name, ilComponent);
+}
