+++ src/components/RelationSearch/RelationSearch.vue
@@ -21,33 +21,23 @@
 
         <h3>Search</h3>
         <button @click='search'>Search</button>
+        <button v-show="relations !== null && !empty_relations"
+                @click="show_search=false">
+          Hide Search
+        </button>
     </div>
     <div v-show='!show_search'>
-      <span id='search-reopen' @click='show_search=true'>Search...</span>
+      <button id='search-reopen' @click='show_search=true'>Edit Search</button>
     </div>
+    <hr>
 
     <div id='result-box' v-show='relations !== null'>
       <h3>Results</h3>
-      <h2 v-show='empty_relations'>Nothing found.</h2>
-      <div class='row' v-for='rel in relations' :key='rel.id'>
-        <div class='col text-left'>
-          <h4 v-html="rel.english"></h4>
-        </div>
-        <div class="col text-right">
-          <span v-for="(src_group, cat, idx) in sources" :key="cat">
-            <span v-if="idx > 0" class="badge badge-source">|</span>
-            <span v-for="src in src_group"
-                  :key="src">
-              <span class='badge badge-source'
-                    v-if="src in rel.source_counts"
-                    :title="src"
-                    :class="srcBadge(src)">
-                {{ rel.source_counts[src] }}
-              </span>
-            </span>
-          </span>
-        </div>
-      </div>
+      <hr>
+      <h4 v-show='empty_relations'>Nothing found.</h4>
+      <span v-for="rel in relations" :key="rel.id">
+        <relation v-bind="rel"></relation>
+      </span>
     </div>
   </div>
 </template>
@@ -126,10 +116,6 @@
         if (this.relations.length > 0)
           this.show_search = false;
       },
-
-      srcBadge: function(src) {
-        return `source-${src}`
-      },
     },
     computed: {
       empty_relations: function() {
@@ -139,10 +125,6 @@
           return true;
         return false
       },
-
-      sources: function() {
-        return this.$sources;
-      }
     },
     created: function() {
       this.addAgent();
