+++ src/components/index.js
@@ -4,3 +4,4 @@ export {default as Evidence} from './Evidence'
 export {default as CurationRow} from './CurationRow'
 export {default as AgentSelect} from './AgentSelect'
 export {default as RelationSearch} from './RelationSearch'
+export {default as Relation} from './Relation'
