+++ src/index.js
@@ -19,3 +19,4 @@ export {default as RefLink} from './components/RefLink'
 export {default as CurationRow} from './components/CurationRow'
 export {default as AgentSelect} from './components/AgentSelect'
 export {default as RelationSearch} from './components/RelationSearch'
+export {default as Relation} from './components/Relation'
